==> 05_attempt/5_training_best_model_4.R <==
# Final Project ----
# Select final model and train it

# load packages ----
library(tidyverse)
library(tidymodels)
library(here)

# loading necessary data
load(here("05_attempt/results/rf_tune_3.rda"))
load(here("05_attempt/data/training_data.rda"))

# Best Model --------
select_best(rf_tune_3, metric = "roc_auc")

# finalize workflow for roc-----
final_wflow_roc <- rf_tune_3 |>
  extract_workflow(rf_tune_3) |>
  finalize_workflow(select_best(rf_tune_3, metric = "roc_auc"))

# train final model----
set.seed(153)
final_fit_rf <-  fit(final_wflow_roc, training_data)

# saving results-------
save(final_fit_rf, file = here("05_attempt/results/final_fit_rf.rda"))
